--- backend/src/main/java/com/devmatch/backend/analysis/domain/AnalysisResult.java
@@ -1,4 +1,0 @@
-package com.devmatch.backend.analysis.domain;
-
-public class AnalysisResult {
-}
