+++ backend/src/main/java/com/devmatch/backend/analysis/entity/AnalysisResult.java
@@ -1,0 +1,4 @@
+package com.devmatch.backend.analysis.entity;
+
+public class AnalysisResult {
+}
