--- backend/src/main/java/com/devmatch/backend/application/domain/Application.java
@@ -1,4 +1,0 @@
-package com.devmatch.backend.application.domain;
-
-public class Application {
-}
