+++ backend/src/main/java/com/devmatch/backend/application/entity/Application.java
@@ -1,0 +1,4 @@
+package com.devmatch.backend.application.entity;
+
+public class Application {
+}
