--- backend/src/main/java/com/devmatch/backend/auth/domain/Auth.java
@@ -1,4 +1,0 @@
-package com.devmatch.backend.auth.domain;
-
-public class Auth {
-}
