+++ backend/src/main/java/com/devmatch/backend/auth/entity/Auth.java
@@ -1,0 +1,4 @@
+package com.devmatch.backend.auth.entity;
+
+public class Auth {
+}
