--- backend/src/main/java/com/devmatch/backend/project/domain/Project.java
@@ -1,4 +1,0 @@
-package com.devmatch.backend.project.domain;
-
-public class Project {
-}
