+++ backend/src/main/java/com/devmatch/backend/project/entity/Project.java
@@ -1,0 +1,4 @@
+package com.devmatch.backend.project.entity;
+
+public class Project {
+}
