--- backend/src/main/java/com/devmatch/backend/user/domain/User.java
@@ -1,4 +1,0 @@
-package com.devmatch.backend.user.domain;
-
-public class User {
-}
