+++ backend/src/main/java/com/devmatch/backend/user/entity/User.java
@@ -1,0 +1,4 @@
+package com.devmatch.backend.user.entity;
+
+public class User {
+}
